added singlenode & cluster folder under named_vectors_tests
