+++ test/acceptance_with_go_client/named_vectors_tests/named_vectors_aggregate.go
