+++ test/acceptance_with_go_client/named_vectors_tests/named_vectors_restart.go
@@ -24,7 +24,7 @@ import (
 	"github.com/weaviate/weaviate/test/docker"
 )
 
-func testRestart(compose *docker.DockerCompose) func(t *testing.T) {
+func TestRestart(compose *docker.DockerCompose) func(t *testing.T) {
 	return func(t *testing.T) {
 		ctx := context.Background()
 		host := compose.GetWeaviate().URI()
